hazman buat
