+++ resources/views/hisdb/apptrsc/apptrsc.blade.php
@@ -1,12 +1,10 @@
-<?php 
-	include_once('../../../../header.php'); 
-?>
+@extends('layouts.main')
 
-<body style="display:none">
+@section('title', 'Appointment')
 
-<input id="Class2" name="Class" type="hidden" value="<?php echo $_GET['TYPE'];?>">
-
+@section('body')
 	<div class='row'>
+		<input id="Class2" name="Type" type="hidden" value="{{Request::get('TYPE')}}">
 		<form id="searchForm" class="formclass" style='width:99%'>
 			<fieldset>
 				<div class='col-md-12' style='padding: 0 0 5px 0;'>
@@ -27,6 +25,14 @@
 				</div>
 			 </fieldset> 
 		</form>
+
+		<div class="panel panel-default">
+    		<div class="panel-body">
+    			<div class='col-md-12' style="padding:0 0 15px 0">
+    				<div id="calendar"></div>
+				</div>
+    		</div>
+		</div>
 
 		<!-- Dialog -->
 		<div id="dialogForm" title="Add Form">
@@ -130,67 +136,6 @@
 		</div>
 
 		<!-- Modal -->
-		<!-- <div class="modal fade" id="ModalAdd" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
-		  <div class="modal-dialog" role="document">
-			<div class="modal-content">
-			<form id="addForm" class="form-horizontal" style="width:99%" method="POST" action="addEvent.php">
-			
-			  <div class="modal-header">
-				<button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-				<h4 class="modal-title" id="myModalLabel">Add Event</h4>
-			  </div>
-			  <div class="modal-body">
-				  <div class="form-group">
-					<label for="title" class="col-sm-2 control-label">Title</label>
-					<div class="col-sm-10">
-					  <input type="text" name="title" class="form-control" id="title" placeholder="Title">
-					</div>
-				  </div>
-					<div class="form-group">
-					<label for="doctor" class="col-sm-2 control-label">Doctor</label>
-					<div class="col-sm-10">
-					  <input type="text" name="doctor" class="form-control" id="doctor" placeholder="Doctor" readonly>
-					</div>
-				  </div>
-				  <div class="form-group">
-					<label for="color" class="col-sm-2 control-label">Color</label>
-					<div class="col-sm-10">
-					  <select name="color" class="form-control" id="color">
-						  <option value="">Choose</option>
-						  <option style="color:#0071c5;" value="#0071c5">&#9724; Dark blue</option>
-						  <option style="color:#40E0D0;" value="#40E0D0">&#9724; Turquoise</option>
-						  <option style="color:#008000;" value="#008000">&#9724; Green</option>						  
-						  <option style="color:#FFD700;" value="#FFD700">&#9724; Yellow</option>
-						  <option style="color:#FF8C00;" value="#FF8C00">&#9724; Orange</option>
-						  <option style="color:#FF0000;" value="#FF0000">&#9724; Red</option>
-						  <option style="color:#000;" value="#000">&#9724; Black</option>
-						</select>
-					</div>
-				  </div>
-				  <div class="form-group">
-					<label for="start" class="col-sm-2 control-label">Start date</label>
-					<div class="col-sm-10">
-					  <input type="text" name="start" class="form-control" id="start" >
-					</div>
-				  </div>
-				  <div class="form-group">
-					<label for="end" class="col-sm-2 control-label">End date</label>
-					<div class="col-sm-10">
-					  <input type="text" name="end" class="form-control" id="end" >
-					</div>
-				  </div>
-				
-			  </div>
-			  <div class="modal-footer">
-				<button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-				<button type="submit" class="btn btn-primary">Save changes</button>
-			  </div>
-			</form>
-			</div>
-		  </div>
-		</div> -->
-
-		<!-- Modal -->
 		<div class="modal fade" id="ModalEdit" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
 		  <div class="modal-dialog" role="document">
 			<div class="modal-content">
@@ -242,204 +187,12 @@
 		  </div>
 		</div>
 
-
-		<!-- JQGRID -->
-		<div class="panel panel-default">
-		    		<div class="panel-body">
-		    			<div class='col-md-12' style="padding:0 0 15px 0">
-            				<div id="calendar"></div>
-        				</div>
-		    		</div>
-		</div>
-
-		<div id="panel" class="panel panel-default">
-		    		<div class="panel-body">
-		    			<div class='col-md-12' style="padding:0 0 15px 0">
-            				<table id="jqGrid" class="table table-striped"></table>
-            					<div id="jqGridPager"></div>
-        				</div>
-		    		</div>
-		</div>
-
- 		<div id="panel2" class="panel panel-default">
-		    <div class="panel-body">
-		    	<div class='col-md-8'>
-            		<table id="detail" class="table table-striped"></table>
-            			<div id="jqGridPager2"></div>
-        		</div>
-
-        		<div class='col-md-4'>
-            		<table id="itemExpiry" class="table table-striped"></table>
-            			<div id="jqGridPager3"></div>
-        		</div>
-		    </div>
-		</div>
-
     </div>
 
-	<?php 
-		include_once('../../../../footer.php');
-	?>
-	<script src="appointment2.js"></script>
-	<script src="../../../../assets/js/utility.js"></script>
-	<script src="../../../../assets/js/dialogHandler.js"></script>
+	@endsection
 
-	<!-- Calendar Script -->
-	<script>
+@section('scripts')
 
-	$(document).ready(function() {
-		
-		$("body").show();
-		var d = new Date();
-
-		$("#dialogForm").dialog({
-			autoOpen: false,
-			width: 9.5 / 10 * $(window).width(),
-			modal: true
-		});
-		
-		$('#calendar').fullCalendar({
-			header: {
-				left: 'prev,next today myCustomButton',
-				center: 'title',
-				right: 'month,agendaWeek,agendaDay'
-			},
-			customButtons: {
-        myCustomButton: {
-            text: 'Add',
-            click: function() {
-                // var temp = $('#resourcecode').val();
-				
-								// $('#ModalAdd #doctor').val(temp);
-								
-								// $('#ModalAdd #start').datetimepicker({
-									// 	format: 'YYYY-MM-DD HH:mm:ss'
-									// }).val(moment(start).format('YYYY-MM-DD HH:mm:ss'));
-									
-									// $('#ModalAdd #end').datetimepicker({
-										// 	format: 'YYYY-MM-DD HH:mm:ss'
-										// }).val(moment(end).format('YYYY-MM-DD HH:mm:ss'));
-										
-										// $('#ModalAdd #start').val(moment(start).format('YYYY-MM-DD HH:mm:ss'));
-										// $('#ModalAdd #end').val(moment(end).format('YYYY-MM-DD HH:mm:ss'));
-
-								 var temp = $('#resourcecode').val();
-								
-								 var start = $(".fc-myCustomButton-button").data( "start");
-								 var end = $(".fc-myCustomButton-button").data( "end");
-
-									$('#dialogForm #doctor').val(temp);
-									
-									$('#dialogForm #start').datetimepicker({
-											format: 'YYYY-MM-DD HH:mm:ss',
-											stepping: 15
-										}).val(moment(start).format('YYYY-MM-DD HH:mm:ss'));
-									
-								$('#dialogForm #end').datetimepicker({
-										format: 'YYYY-MM-DD HH:mm:ss',
-										stepping: 15
-									}).val(moment(end).format('YYYY-MM-DD HH:mm:ss'));
-									
-									$('#dialogForm #start').val(moment(start).format('YYYY-MM-DD HH:mm:ss'));
-									$('#dialogForm #end').val(moment(end).format('YYYY-MM-DD HH:mm:ss'));
-									
-									$("#dialogForm").dialog("open");
-									// $('#ModalAdd').modal('show');
-            }
-        }
-    },
-			// defaultDate: '2016-01-12',
-			defaultDate: d,
-			navLinks: true, // can click day/week names to navigate views
-			editable: true,
-			eventLimit: true, // allow "more" link when too many events
-			selectable: true,
-			selectHelper: true,
-			select: function(start, end) {
-				$('#calendar').fullCalendar('changeView', 'agendaDay', moment(start).format('YYYY-MM-DD'));
-				$(".fc-myCustomButton-button").data( "start", start );
-				$(".fc-myCustomButton-button").data( "end", end );
-			},
-			eventRender: function(event, element) {
-				element.bind('dblclick', function() {
-					$('#ModalEdit #id').val(event.id);
-					$('#ModalEdit #title').val(event.title);
-					$('#ModalEdit #color').val(event.color);
-					$('#ModalEdit').modal('show');
-				});
-			},
-			eventDrop: function(event, delta, revertFunc) { // si changement de position
-
-				edit(event);
-
-			},
-			eventResize: function(event,dayDelta,minuteDelta,revertFunc) { // si changement de longueur
-
-				edit(event);
-
-			},
-			events: {
-				url:'getEvent.php',
-				type:'POST',
-				data:{
-					drrsc:''
-				}
-			}
-			
-		});
-		
-		function edit(event){
-			start = event.start.format('YYYY-MM-DD HH:mm:ss');
-			if(event.end){
-				end = event.end.format('YYYY-MM-DD HH:mm:ss');
-			}else{
-				end = start;
-			}
-			
-			id =  event.id;
-			
-			Event = [];
-			Event[0] = id;
-			Event[1] = start;
-			Event[2] = end;
-			
-			$.ajax({
-			 url: 'editEventDate.php',
-			 type: "POST",
-			 data: {Event:Event},
-			 success: function(rep) {
-					if(rep == 'OK'){
-						// alert('Saved');
-					}else{
-						alert('Could not be saved. try again.'); 
-					}
-				}
-			});
-		}
-		
-	});
-
-	$('#submit').click(function(){
-		$.post("addEvent.php", $("#addForm").serialize(), function (data) {
-		}).fail(function (data) {
-			//////////////////errorText(dialog,data.responseText);
-		}).done(function (data) {
-			$("#dialogForm").dialog('close');
-			var events = {
-							url: "getEvent.php",
-							type: 'POST',
-							data: {
-								drrsc: $('#resourcecode').val()
-							}
-						}
-				
-			$('#calendar').fullCalendar( 'removeEventSource', events);
-			$('#calendar').fullCalendar( 'addEventSource', events);         
-			$('#calendar').fullCalendar( 'refetchEvents' );
-		});
-	});
-
-</script>
-
-</body>
-</html>
+	<script src="js/hisdb/apptrsc/apptrsc.js"></script>
+	
+@endsection
